Implement TotalPrice of ShoppingCart. Encapsulate adding product to shopping cart in WebUI ShoppingCartService.

+++ ShoppingCartAPI/Entities/ShoppingCartItem.cs
@@ -13,7 +13,7 @@ namespace ShoppingCartAPI.Entities
         public Guid ProductId { get; init; }
 
         public string? ProductName { get; init; }
-        public double ProductPrice { get; init; }
+        public decimal ProductPrice { get; init; }
         public ushort Quantity { get; init; }
     }
 }
